backend/api/requirements.txt
# FastAPI==0.104.1
# uvicorn==0.24.0
# pydantic==2.5.0
# python-dotenv==1.1.1
# requests==2.32.5
# sqlalchemy==2.0.23
# psycopg2-binary==2.9.9
# alembic==1.13.0
# python-multipart==0.0.6
# passlib==1.7.4
# bcrypt==4.1.2
# python-jose==3.3.0
# cryptography==41.0.8

# Core FastAPI and server
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0

# Environment and configuration
python-dotenv>=1.0.0,<2.0.0

# HTTP client for external APIs
requests>=2.32.0,<3.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
alembic>=1.13.0,<2.0.0

# File handling and forms
python-multipart>=0.0.6,<1.0.0

# Authentication and security
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.1.0,<5.0.0
python-jose[cryptography]>=3.3.0,<4.0.0

# Additional utilities
email-validator>=2.0.0,<3.0.0
Pillow>=10.0.0,<11.0.0

# Development and testing (optional)
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0
httpx>=0.25.0,<1.0.0

# Data processing (for ML/AI features)
pandas>=2.1.0,<3.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0

# For async database operations
asyncpg>=0.29.0,<1.0.0

# CORS and middleware
python-cors>=1.0.0,<2.0.0

# Logging and monitoring
structlog>=23.2.0,<24.0.0

# Background tasks
celery>=5.3.0,<6.0.0
redis>=5.0.0,<6.0.0